+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All the notable changes in the project would be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [Unreleased](https://github.com/ArindomGhosh/stategenie/compare/v1.0.0...main)
+## [Unreleased](https://github.com/ArindomGhosh/stategenie/compare/v1.1.0...main)
+
+## [1.1.0](https://github.com/ArindomGhosh/stategenie/compare/v1.0.0...v1.1.0)
+### Added
+- DataStateGenie: Generate ui-state data class with update extention function.
+- wiki: Wiki pages updated to ease the readability.
+
+### Modified
+- spotless: modified to apply only on the diff from main.
 
 ## [1.0.0](https://github.com/ArindomGhosh/stategenie/compare/v0.2.1...v1.0.0)
 ### Added